+++ Source/DelayEngine.cpp
@@ -1,6 +1,5 @@
 #include "DelayEngine.h"
 #include "DSP.h"
-#include "Types.h"
 
 DelayEngine::DelayEngine()
 {
@@ -19,23 +18,23 @@ void DelayEngine::prepareToPlay(double sampleRate, int samplesPerBlock) noexcept
 	spec.numChannels = 2;
 	m_sampleRate = static_cast<float>(sampleRate);
 	m_feedbackCompressor.prepare(spec);
+	m_insertEffectSelector.prepareToPlay(sampleRate, samplesPerBlock);
 	m_filterEngine.prepareToPlay(sampleRate, samplesPerBlock);
 	m_feedbackHighpass.prepare(spec);
-	m_chorusEngine.prepareToPlay(sampleRate, samplesPerBlock);
 	m_stereoDelay.prepareToPlay(sampleRate, samplesPerBlock);
-	m_coeff = onePoleLowpassCoeff(100.0f, static_cast<float>(sampleRate));
-
+	m_coeff = onePoleLowpassCoeff(100.0f, m_sampleRate);
+	m_killCoeff = onePoleLowpassCoeff(10.0f, m_sampleRate);
 }
 
 void DelayEngine::reset() noexcept
 {
 	m_stereoDelay.reset();
-	m_chorusEngine.reset();
 	m_filterEngine.reset();
 	m_feedbackHighpass.reset();
 	m_feedbackHighpass.setCutoffFrequency(60.0f); // just to remove rumble from fb loop
 	m_feedbackHighpass.setResonance(Parameters::defaultFilterQ);
 	m_feedbackCompressor.reset();
+	m_insertEffectSelector.reset();
 	m_feedbackL = 0.0f;
 	m_feedbackR = 0.0f;
 	m_baseDelayTimeMs = -1.0f;
@@ -52,6 +51,15 @@ void DelayEngine::setDelayMode(const int modeIndex)
 	}
 }
 
+void DelayEngine::setInsertEffect(const int effectIndex)
+{
+	InsertEffectType newEffect = static_cast<InsertEffectType>(effectIndex);
+	if (newEffect != m_currentInsertEffect) {
+		m_currentInsertEffect = newEffect;
+		m_insertEffectSelector.select(m_currentInsertEffect);
+	}
+}
+
 void DelayEngine::setDelayTimes(const float targetL, const float targetR) {
 	float targetLeftMs = targetL - m_offsetMs;
 	float targetRightMs = targetR + m_offsetMs;
@@ -59,6 +67,8 @@ void DelayEngine::setDelayTimes(const float targetL, const float targetR) {
 	if (std::abs(m_delayTimeMsR) < 1e-4f) m_delayTimeMsR = targetRightMs;
 	m_delayTimeMsL = onePoleLowpass(targetLeftMs, m_delayTimeMsL, m_coeff);
 	m_delayTimeMsR = onePoleLowpass(targetRightMs, m_delayTimeMsR, m_coeff);
+	if (std::abs(m_delayTimeMsL - targetLeftMs) < 0.01f) m_delayTimeMsL = targetLeftMs;
+	if (std::abs(m_delayTimeMsR - targetRightMs) < 0.01f) m_delayTimeMsR = targetRightMs;
 	float leftMs = juce::jlimit(Parameters::minDelayTime, Parameters::maxDelayTime, m_delayTimeMsL);
 	float rightMs = juce::jlimit(Parameters::minDelayTime, Parameters::maxDelayTime, m_delayTimeMsR);
 	m_stereoDelay.setDelayTime(leftMs, Channel::Left);
@@ -73,7 +83,35 @@ void DelayEngine::update(const Parameters& params)
 	setWidthLevel(params.stereo());
 	setOffsetMs(params.offset());
 	setDelayTimes(params.delayTimeL(), params.delayTimeR());
+	setKill(params.feedbackKill());
 	m_filterEngine.update(params);
+	m_insertEffectSelector.update(params);
+}
+
+void DelayEngine::setKill(bool shouldKill)
+{
+	bool risingEdge = shouldKill && !m_killButtonHeld;
+	m_killButtonHeld = shouldKill;
+	if (risingEdge)
+		triggerKill();
+}
+
+void DelayEngine::triggerKill()
+{
+	m_isKilling = true;
+}
+
+void DelayEngine::resetAfterKill()
+{
+	m_stereoDelay.kill();
+	m_insertEffectSelector.kill();
+	m_filterEngine.kill();
+	m_feedbackHighpass.reset();
+	m_feedbackCompressor.reset();
+	m_feedbackL = 0.0f;
+	m_feedbackR = 0.0f;
+	m_killGain = 1.0f;
+	m_isKilling = false;
 }
 
 void DelayEngine::updateSyncedModulators(const double ppqPosition)
@@ -81,10 +119,16 @@ void DelayEngine::updateSyncedModulators(const double ppqPosition)
 	m_filterEngine.updateLFOs(ppqPosition);
 }
 
-void DelayEngine::processSample(const float& inL, const float& inR, float& outL, float& outR, const Parameters& params)
+void DelayEngine::processSample(const float& inL, const float& inR, float& outL, float& outR)
 {
 	float dryL = inL;
 	float dryR = inR;
+
+	if (m_isKilling) {
+		m_killGain = onePoleLowpass(0.0f, m_killGain, m_killCoeff);
+		if (m_killGain < 0.001f)
+			resetAfterKill();
+	}
 
 	float mono = (dryL + dryR) * .5f;
 	float delayInL = 0.0f, delayInR = 0.0f, wetL = 0.0f, wetR = 0.0f;
@@ -97,7 +141,7 @@ void DelayEngine::processSample(const float& inL, const float& inR, float& outL,
 			delayInL = m_feedbackR;
 			delayInR = mono + m_feedbackL;
 			break;
-		case DelayMode::Cross: 
+		case DelayMode::Cross:
 			delayInL = dryL + m_feedbackR;
 			delayInR = dryR + m_feedbackL;
 			break;
@@ -112,8 +156,9 @@ void DelayEngine::processSample(const float& inL, const float& inR, float& outL,
 
 	m_stereoDelay.processSample(delayInL, wetL, Channel::Left);
 	m_stereoDelay.processSample(delayInR, wetR, Channel::Right);
-	
-	m_chorusEngine.processSample(wetL, wetR, wetL, wetR, params);
+	wetL *= m_killGain;
+	wetR *= m_killGain;
+	m_insertEffectSelector.processSample(wetL, wetR, wetL, wetR);
 
 	m_filterEngine.processSample(wetL, wetR, wetL, wetR);